rearrange/split files
