+++ 3_email/Rakefile
@@ -1,0 +1,5 @@
+require 'rake/testtask'
+
+Rake::TestTask.new do |t|
+  t.pattern = "spec/verifiers/*_spec.rb"
+end
